% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{append_bbox_info}
\alias{append_bbox_info}
\alias{filter_on_bbox}
\title{Create a bounding box for each geometries in an sf object}
\usage{
append_bbox_info(sf_obj)

filter_on_bbox(aoi, sf_obj)
}
\arguments{
\item{sf_obj}{sf object}

\item{aoi}{geometry of area of interest}
}
\value{
same sf object with 4 new columns : xmin, ymin, xmax and ymax

sf object filtered on area of interest
}
\description{
Create a bounding box for each geometries in an sf object

filter sf object on an area of interest using bounding box information created by \code{append_bbox_info}
}
\details{
Used for faster lookup when selecting area of interest.
}
