% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffers}
\alias{buffers}
\title{Compute buffers around geometries}
\usage{
buffers(sf_obj, dist)
}
\arguments{
\item{sf_obj}{A spatial feature object from which geometries can be extracted.}

\item{dist}{A named integer vector. buffer distances as produced by \code{distances}.}
}
\value{
A list of named buffers as spatial feature objects. Names recycled from \code{dist}.
}
\description{
Compute buffers around geometries
}
