% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_donut.R
\name{compute_donut}
\alias{compute_donut}
\title{Create a buffer geometry of distance \code{dist} around existing geometries}
\usage{
compute_donut(dist, sf_obj)
}
\arguments{
\item{dist}{distance of buffer}

\item{sf_obj}{sf object}
}
\value{
list
}
\description{
Create a buffer geometry of distance \code{dist} around existing geometries
}
