% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_in_zone.R
\name{compute_in_zone}
\alias{compute_in_zone}
\title{compute time in incursion zones and creates a summary table}
\usage{
compute_in_zone(loi, incz)
}
\arguments{
\item{loi}{geometry linestring of interest}

\item{incz}{geometries of incursion zones}
}
\value{
list summary table of time and segments in incursion zones
}
\description{
compute time in incursion zones and creates a summary table
}
