% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_dist}
\alias{default_dist}
\alias{default_zones}
\title{Default buffer distances (low = 1500m, moderate = 1000m, high = 500m) using reference label
\code{in_UWR} created from a call to \code{distances}.}
\usage{
default_dist()

default_zones(recache = FALSE)
}
\arguments{
\item{recache}{A boolean. Should the legal habitat area zones be retrieved from source again and
resaved again into the cache directory. Default to FALSE.}
}
\description{
Default buffer distances (low = 1500m, moderate = 1000m, high = 500m) using reference label
\code{in_UWR} created from a call to \code{distances}.

Default legal habitat area zones as defined in the document
\verb{Habitat Base Layers for Flight Monitoring Application.docx} from the github repository.
}
