% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_results.R
\name{empty_results}
\alias{empty_results}
\title{create template data.table for flight summary results}
\usage{
empty_results(flight, dist)
}
\arguments{
\item{flight}{sf object of flight data}

\item{dist}{numeric distance}
}
\value{
data.table object
}
\description{
create template data.table for flight summary results
}
