% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export analysis results}
\usage{
export(x, file, flight_id = NULL)
}
\arguments{
\item{x}{The result object to be written.}

\item{file}{A filename using kml extension.}

\item{flight_id}{Flight ID for filtering.}
}
\description{
Export analysis results
}
