% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_on_bbox.R
\name{filter_on_bbox}
\alias{filter_on_bbox}
\title{filter sf object on an area of interest using bounding box information created by \code{append_bbox_info}}
\usage{
filter_on_bbox(aoi, sf_obj)
}
\arguments{
\item{aoi}{geometry of area of interest}

\item{sf_obj}{sf object}
}
\value{
sf object filtered on area of interest
}
\description{
filter sf object on an area of interest using bounding box information created by \code{append_bbox_info}
}
