% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered.R
\name{filtered}
\alias{filtered}
\title{Compute masked points index}
\usage{
filtered(
  sf_obj,
  zone,
  aoi,
  max_dist,
  max_altitude,
  viewshed = TRUE,
  check_tiles = FALSE
)
}
\arguments{
\item{sf_obj}{A spatial feature object of flight track points.}

\item{zone}{A geometry of type \code{sfc_POLYGON} representing the reference zone without buffers.}

\item{aoi}{Area of interest used to extract the right digital elevation tiles.}

\item{max_dist}{Maximum distance below which a track point can be considered.}

\item{max_altitude}{distance in meters to represent the maximum altitude for a track point to be considered.}

\item{viewshed}{A logical. Perform viewshed analysis to remove flight segments masked by terrain.
Default to TRUE. Most compute intensive task of the analysis.}

\item{check_tiles}{Should the tiles that you already have in your cache be checked to see if
they need updating? Default FALSE. If you are running the same code frequently and are confident
the tiles haven't changed, setting this to FALSE will speed things up. For digital elevation
model from \code{bcmaps::cded}.}
}
\value{
A logical vector.
}
\description{
Compute masked points index
}
