% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight_process.R
\name{flight_process}
\alias{flight_process}
\title{process flight data to produce summary table of segments in zones of interest}
\usage{
flight_process(
  flight,
  zones,
  telemetry,
  dist,
  max_altitude = units::set_units(500, m),
  segments = TRUE
)
}
\arguments{
\item{flight}{flight data obtained using \code{read_GPX}}

\item{zones}{geometries of incursion zones}

\item{telemetry}{geometries of points of interest using telemetry information}

\item{dist}{list named list of distance in meters defined using \code{set_units}}

\item{max_altitude}{distance in meters defined using \code{set_units} to represent the maximum altitude for a track point to be considered}

\item{segments}{boolean, if TRUE segments within zones of interest are returned. Defaults to TRUE}
}
\value{
list summary table of time and segments in incursion zones
}
\description{
process flight data to produce summary table of segments in zones of interest
}
