% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.flightanalysis}
\alias{plot.flightanalysis}
\title{Print flight analysis}
\usage{
\method{plot}{flightanalysis}(
  x,
  y,
  ...,
  backend = c("leaflet", "ggplot2"),
  flight_id = NULL,
  proxy = NULL
)
}
\arguments{
\item{x}{object}

\item{y}{ignored}

\item{...}{further specifications.}

\item{backend}{A character string. Either \code{leaflet} or \code{ggplot2}.}

\item{flight_id}{An integer. Which flight ids to plot. Default to all.}

\item{proxy}{For \code{leaflet} backend, a \code{leaflet::leafletProxy} object can be passed instead of initializing a new one.}
}
\description{
Print flight analysis
}
