% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.flightanalysis}
\alias{print.flightanalysis}
\title{Print flight analysis}
\usage{
\method{print}{flightanalysis}(x, ...)
}
\arguments{
\item{x}{Flight analysis results to print}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print flight analysis
}
