% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_flight.R
\name{process_flight}
\alias{process_flight}
\title{Process flight data and produce summary table of time in zones}
\usage{
process_flight(
  flight,
  zones = default_zones(),
  dist = default_dist(),
  max_altitude = 500,
  geom_out = TRUE,
  check_tiles = FALSE
)
}
\arguments{
\item{flight}{Flight data obtained from subsetting results of \code{ingest_flights}.}

\item{zones}{Geometries of incursion zones and telemetries. Default to \code{default_zones()}.}

\item{dist}{A named numeric vector of distances from \code{distances}. Default to \code{default_dist()}.}

\item{max_altitude}{Maximum altitude in meters for a track point to be considered.}

\item{geom_out}{A logical. If TRUE, some geometries used for the analysis are returned.
Defaults to TRUE.}

\item{check_tiles}{Should the tiles that you already have in your cache be checked to see if
they need updating? Default FALSE. If you are running the same code frequently and are confident
the tiles haven't changed, setting this to FALSE will speed things up. For digital elevation
model from \code{bcmaps::cded}.}
}
\value{
A list with a summary table of time in zones and optional geometries.
}
\description{
Process flight data and produce summary table of time in zones
}
