% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flight.R
\name{read_flight}
\alias{read_flight}
\alias{read_flight_file}
\alias{read_GPX}
\alias{read_KML}
\alias{read_flight_db}
\title{Ingest flights data source}
\usage{
read_flight(dsn, query = NULL)

read_flight_file(fpath)

read_GPX(fpath)

read_KML(fpath)

read_flight_db(dsn, query)
}
\arguments{
\item{dsn}{data source name. Either a character vector of file paths of a connection to a
a database. When a connection is used, it needs a query to retrieve flight track records.}

\item{query}{SQL query to select records.}

\item{fpath}{A file path.}
}
\value{
A list of flights each represented by a list of spatial feature objects
}
\description{
Ingest flights data source

Read a flight file

Read a GPX file into a list of simple feature objects

Read a KML file into a list of simple feature object

Read flights from TRACKS database
}
