% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_GPX.R
\name{read_GPX}
\alias{read_GPX}
\title{Read a GPX file into a list of simple feature object}
\usage{
read_GPX(file)
}
\arguments{
\item{file}{A character. Path to a GPX file.}
}
\value{
sf object
}
\description{
Read a GPX file into a list of simple feature object
}
