% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_KML.R
\name{read_KML}
\alias{read_KML}
\title{Read a KML file into a list of simple feature object}
\usage{
read_KML(file)
}
\arguments{
\item{file}{character string of filename}
}
\value{
sf object
}
\description{
Read a KML file into a list of simple feature object
}
