% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_contiguous_duplicates.R
\name{remove_contiguous_duplicates}
\alias{remove_contiguous_duplicates}
\title{create template data.table for flight summary results}
\usage{
remove_contiguous_duplicates(tp, keep)
}
\arguments{
\item{tp}{sf object of track points layer of flight data}

\item{keep}{character vector of column to keep}
}
\value{
sf object
}
\description{
create template data.table for flight summary results
}
