% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_in_zone.R
\name{time_in_zone}
\alias{time_in_zone}
\title{Compute time spent in zones}
\usage{
time_in_zone(sf_obj, zones, exclude = c("Buffers", "All"))
}
\arguments{
\item{sf_obj}{A spatial feature object with \code{LINESTRING} geometries and field \code{time_deltas}.}

\item{zones}{A named list of \code{POLYGON} geometries.}

\item{exclude}{Character vector of zone names to exclude.}
}
\value{
A list with \code{LINESTRING} segments and time spent in each zone.
}
\description{
Compute time spent in zones
}
\details{
Also return an \code{All}
element which is the sum of of time spent in each zones. It is not the time spent in a union
of all zones geometries. If the zones overlap, this information would be important to
communicate to avoid confusion.
}
