% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_bbox_info.R
\name{append_bbox_info}
\alias{append_bbox_info}
\title{Create a bounding box for each geometries in an sf object}
\usage{
append_bbox_info(sf_obj)
}
\arguments{
\item{sf_obj}{sf object}
}
\value{
sf object with 4 new columns : xmin, ymin, xmax and ymax
}
\description{
Create a bounding box for each geometries in an sf object
}
