% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{distances}
\alias{distances}
\title{Define buffer distances}
\usage{
distances(..., reflabel = "In UWR")
}
\arguments{
\item{...}{Named buffer distances. If unnamed, will use the distance as names.}

\item{reflabel}{A character string. Reference geometries name from which the
buffers are computed. Default to \verb{In UWR}.}
}
\value{
A sorted named integer vectors.
}
\description{
Define buffer distances
}
\examples{
distances(High = 1500, Moderate = 1000, Low = 500, reflabel = "In UWR")
}
