% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_GPX.R
\name{read_GPX}
\alias{read_GPX}
\title{read a gpx file into an r object}
\usage{
read_GPX(file)
}
\arguments{
\item{file}{character string of filename}
}
\value{
sf object
}
\description{
read a gpx file into an r object
}
