% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_KML.R
\name{read_KML}
\alias{read_KML}
\title{read a kml file into an r object}
\usage{
read_KML(file)
}
\arguments{
\item{file}{character string of filename}
}
\value{
sf object
}
\description{
read a kml file into an r object
}
